changed "programmed C++" which wasn't grammatical to "written C++", which also sets up a nice parallel to the newspaper.
I guess my concern was that it would look like I hadn't written anything
but C++. We'll see if this survives more revisions.

Changed "you had better do it well" to "it had better be done well" to
keep "you" out of it.

Shortened "I love to write" since I edit too, and there's a lot of the verb
"write" nearby. Also the sentence moves a little faster, which I also
wanted.

Changed "we simulated chip designs" to "we simulated hardware designs"
to make the zinger clearer.