In the vision-doc section, thought the informal "editing" might be less stuffy, since the word before it is "my".
In the L2 section, saw some words that could be struck out. Also
wanted to find the document word count --it's 10K. Some I could see was
Martin's, so I cut 2K off my contribution.

Reading the spec I realized it was not doing just one novel thing,
it was filled with novel things. So I changed "something that had never"
to "a number of things that had never."